requirements.txt
flask~=3.1.0
instaloader~=4.14
pyotp~=2.9.0
waitress~=3.0.2
google-generativeai>=0.7.2
